+++ test/src/client_test.dart
@@ -19,11 +19,8 @@ class MockTusClient extends TusClient {
     Map<String, String>? metadata,
     int maxChunkSize = 512 * 1024,
   }) : super(
-          url,
           file,
           store: store,
-          headers: headers,
-          metadata: metadata,
           maxChunkSize: maxChunkSize,
         ) {
     httpClient = MockClient();
@@ -231,7 +228,10 @@ main() {
     bool success = false;
     double? progress;
     await client.upload(
-        onComplete: () => success = true, onProgress: (p, e, c) => progress = p);
+      onComplete: () => success = true,
+      onProgress: (p, e, c) => progress = p,
+      uri: Uri(),
+    );
 
     expect(success, isTrue);
     expect(progress, equals(100));
@@ -274,6 +274,7 @@ main() {
     await client.upload(
       onComplete: () => success = true,
       onProgress: (p, e, c) => progress = p,
+      uri: Uri(),
     );
 
     expect(success, isTrue);
@@ -317,7 +318,10 @@ main() {
     bool success = false;
     double? progress;
     await client.upload(
-        onComplete: () => success = true, onProgress: (p, e, c) => progress = p);
+      onComplete: () => success = true,
+      onProgress: (p, e, c) => progress = p,
+      uri: Uri(),
+    );
 
     expect(success, isFalse);
     expect(progress, equals(50));
@@ -343,7 +347,10 @@ main() {
     bool success = false;
     double? progress;
     await client.upload(
-        onComplete: () => success = true, onProgress: (p, e, c) => progress = p);
+      onComplete: () => success = true,
+      onProgress: (p, e, c) => progress = p,
+      uri: Uri(),
+    );
 
     expect(success, isTrue);
     expect(progress, equals(100));
@@ -358,7 +365,9 @@ main() {
         .thenAnswer((_) async => http.Response("500 Server Error", 500));
 
     expectLater(
-        () => client.upload(),
+        () => client.upload(
+              uri: Uri(),
+            ),
         throwsA(predicate((e) =>
             e is ProtocolException &&
             e.message ==
@@ -377,7 +386,9 @@ main() {
             http.Response("", 204, headers: {"upload-offset": ""}));
 
     expectLater(
-        () => client.upload(),
+        () => client.upload(
+              uri: Uri(),
+            ),
         throwsA(predicate((e) =>
             e is ProtocolException &&
             e.message ==
@@ -398,7 +409,9 @@ main() {
     )).thenAnswer((_) async => http.Response("500 Server Error", 500));
 
     expectLater(
-        () => client.upload(),
+        () => client.upload(
+              uri: Uri(),
+            ),
         throwsA(predicate((e) =>
             e is ProtocolException &&
             e.message ==
@@ -419,7 +432,9 @@ main() {
     )).thenAnswer((_) async => http.Response("", 204));
 
     expectLater(
-        () => client.upload(),
+        () => client.upload(
+              uri: Uri(),
+            ),
         throwsA(predicate((e) =>
             e is ProtocolException &&
             e.message ==
@@ -441,7 +456,9 @@ main() {
         (_) async => http.Response("", 204, headers: {"upload-offset": "50"}));
 
     expectLater(
-        () => client.upload(),
+        () => client.upload(
+              uri: Uri(),
+            ),
         throwsA(predicate((e) =>
             e is ProtocolException &&
             e.message ==